Start work on F2 lock manager
Intention is to build a lock manager with a fast path to
acquire locks that can be acquired, and a regular wait graph
for waiting on locks.

This aught to be useful for applications that do a lot of shared
locking, but that when an exclusive lock is held can expect to
wait a bit (eg. some IO operation will likely need to happen before
the lock becomes available).