Hashing out more structure: Partitions
F1 uses ConcurrentHashMap and then spinlocks on every lock
to acquire it; I want to try a significantly simpler approach of
grouping locks into partitions, and then doing all the work of
looking a lock up, inserting it into the hashtable if it doesn't exist,
and attempting to acquire it all under the one partition lock.

Trying to articulate but riding in the passenger seat through this
insane town in the Ozarks; some kid named "Preslee" is having her
birthday party at the local BBQ shack today, seems baller.