Introduce first iteration on deadlock detector
This is heavily influenced by the PG deadlock detector; note that it
does not keep a "seen" set; it simply traverses until it can't find
any more edges or until it finds the source it started at.

Intuitively, this seems like it could get stuck in an infinite loop
if there are two deadlocks happening at the same time; it could end
up infinitely circling the second deadlock loop without exiting; or no?

Maybe that's guarded by the fact that there are no forks in the graph;
everything waits on at most one lock?

Def. needs more investigation.