Resolve outstanding issues for combination benchmark
There were several issues uncovered by this benchmark,
both the prior commit that introduced a seen list to the
deadlock detection, as well as the major fix in this commit
which removed an issue where a client would signal itself
when it removed itself from a lock wait list.

The first iteration of the combination benchmark is now complete
and shows several interesting things.

First, F2 already substantially outperforms Forseti, successfully
completing about an order of magnitude more scenarios per ms:

    F2Locks_PerfTest.f2SchemaExclusiveCombintation       thrpt    5  22.211 ± 107.533  ops/ms
    F2Locks_PerfTest.forsetiSchemaExclusiveCombintation  thrpt    5   6.087 ±   3.352  ops/ms

However, as you can also see, the variance for F2 is much larger.
I *think* this has to do with the deadlock check wait period - eg.
F2 will wait up to a second to check for a deadlock currently.

There are several interesting options to explore to improve this,
for now I think the benchmark at least shows the F2 approach overall
shows a lot of promise.