Add internal support for upgrade locking
Modeled by introducing a second lock entry that exists alongside
the SHARED entry a client would already hold; the second entry is
marked UPGRADE and can sit on wait lists just like any other, and
is allowed to be assigned to the exclusiveHolder slot.

There's a good bit of repetition happening now, we may want to refactor
out some commonalities.

Also note that this feature expands the number of possible deadlock
scenarios that could be avoided using wait list reordering. The most
obvious case is guarded for (adding an upgrade request after an
exclusive request).